:bug: No longer modify object values in watchPublic
Previously, to keep track of the paths to children properties to be
watched, a `__watchPublicPath` property was set on all object values,
recursively. One side effect of mutating the value object (which can
affect the outside world) is that public properties that are set to
keyframes are no longer recognized as keyframes, because
`__watchPublicPath` is a key that is not a number.

Now, an internal weak map is used to keep track of the paths to children
properties, so they can be easily tracked while not affecting the
outside world.