:bug: Fix Video dimension issues
Both `width` and `mediaWidth` were being set to the `width` constructor
option, making the `mediaWidth` option useless. Now, they are initiated
independently.

Another related problem was that, if `width`, `clipWidth`, or
`mediaWidth` wasn't specified with a constructor option, it was
initiated to the video's width. Now, that is only true for `width`;
`clipWidth` falls back to `width` and `mediaWidth` falls back to
`clipWidth` *in doRender*. That way, if `width` is changed by the user,
and `clipWidth` isn't set, `clipWidth` will follow, etc. Similarly,
`width` falls back to the movie's width (in `doRender`).