:wrench: Use firefox for running tests with karma
For some reason, running tests that involved recording a movie with an
audio layer resulted in an empty blob when running them alongside the
rest of the test suite. However, when running one of these tests by
itself (in chrome), it produced a non-empty blob. Additionally,
recording a movie with an audio layer worked when opening the 'export'
example in chrome (the same version of chrome).

Running the same tests (along with the rest of the test suite) in
firefox results in non-empty blobs. Since this worked in firefox but not
chrome, I configured Karma to use firefox to run the test suite.

Fixes #100