:rewind: Remove views
This was a tough decision to make, but after seeing how views performed
compared to the old canvas system, it was decided that the benefits of
views were not worth the added complexity that users would face while
using it.

Views only offered an fps boost when a lot of shader effects were used
at once. When tested with a movie with many layers without any effects,
views brought the fps down. Likewise, when tested with a movie with one
layer that had many non-shader effects, views degraded the performance.
See examples/performance/view-fps (deleted in this commit) to compare
the performance of views in your browser.

Now, there were reasons for adding views other than offering a
performance boost. Views did allow for over 16 shader fragments to be
used at once, and they offered a more streamline experience for
developing effects that use graphics APIs like WebGL. However, if we use
a simpler approach to rendering in the future - say, choosing WebGPU as
the only supported rendering engine - we could overcome these obstacles
just as easily, without overcomplicating things.