:rewind: Record on a background canvas once again
When making the movie record to the same canvas that it plays to, I did
not realize that there does not seem to be a way for the movie to record
to the same audio destination as it plays to; recording and outputting
audio at the same time is not possible. For consistency reasons, we
should record to a background canvas, since we record to a background
audio destination.

This reverts commit 0ff101acb8ee470465c0a0b77aa5a02d3e3c5253.