:sparkles: Replace 'ready` event with `whenReady`
This is a much cleaner and safer way to wait for all resources to load.

Before, the layers and effects pushed their ready states to the movie.
To do this, they emitted the 'ready' event every time their ready states
changed. Since the ready state is accessed with a getter (the boolean
`ready` getter), every place in the code that *could* change `ready`
needed to emit the event (when needed).

Now, the movie pulls the ready state from its layers and effects. This
is as simple as wrapping the DOm events in promises and waiting for them
to load.

See #188