:white_check_mark: Throw error and unhandledrejection events
For some reason, Karma doesn't capture errors thrown by the browser.
They show up in the browser but not in the terminal. Errors thrown in
the JavaScript code show up in both the browser console and the
terminal.

I tried updating a lot of options in the Karma config, but nothing
worked. Listening to these events and throwing them as errors fixed
this.