:bug: Fix audio destination reconnection logic
Before this commit, `Video` and `Audio` layers are silent after
recording once. Playing, streaming or recording again does not result in
any sound.

This fix introduces a new error message when updating
examples/introduction/audio.html to record instead of play, but Etro's
support for users directly modifying the audio graph is already very
buggy. This buggy behavior can be illustrated when updating the same
audio example to record instead of play and setting `muter.gain.value`
to a positive value. When making these changes before this commit, the
`layer1` is still muted. It probably is not routed to the destination
node properly.

Closes: #106