:wrench: Replace `browser` with `main` in `package.json`
This allows the package to be resolved in Node.

Lazy-loading etro in the frontend code of a NextJS or SvelteKit project
previously resulted in a compile-time module resolution error:

```
Module not found: Can't resolve 'etro'
```

To use any client-only library in a NextJS project, the package needs
to get resolved in Node and lazily imported in the browser.

Previously, The backend code of a NextJS project could not resolve
etro, which was tagged as client-only (the `browser` option was set).
Replacing the `browser` field with `main` lets the package get imported
in Node, without changing the project's entrypoint.

Closes #243